// frontend/src/services/OpenApi/models/TenureUpdateItemDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type TenureUpdateItemDto = {
    cboaId?: number;
    revisionCount?: number;
    fileId: string;
    cuttingPermit?: string;
    cutBlock: string;
    isPrimary?: boolean;
};